[Refactor] Refactor building system and adding compiled products (#261)
# Refactor building system and adding compiled products

Now, we can build `Function`, `IRModule`, `Task` and `FlowGraph` to
`CompiledFunction`, `CompiledModule`, `CompiledTask` and
`CompiledModule`. The compiled products can be saved on the disk and
loaded for running.

The `CompiledTask` has a `lib.so` that contains all the scheduled
kernels. So we will not have the problem of loading thousands of dynamic
libraries when we support dynamic shape & kernel tuning (e.g., 100
operators * 200 schedules for each operator).

The `CompiledGraph` contains a C runtime that invokles the kernels and
allocates memory. With this, we do not need CUDA graph to get rid of the
framework overhead when inference. The `CompiledGraph` can be stored as
a single file on the disk. Thus, we can compile one model on one machine
and then tune & run the model on another machine.

# Minors

- Add `target` argument to `build_task` and `build_ir_module` to specify
the target for code generation and underlying compiler (e.g., nvcc vs.
gcc)
- Move `IRModule` from `hidet.ir.func` to `hidet.ir.module`.
- Changed the format to define tuning space (may change again in the
future).