[Operators] preliminary symmetric weight quantization (#298)
- Added preliminary symmetric weight quantization in tasks
- Changed Operator._run to Operator.symbolic_run, such that construction
of the operator does not automatically run imperative_run, but
symbolic_run. This is so that constant folding does not happen with
barrier op, and leads to the intended behaviour with quantization
layers. This should be fine to do since all imperative calls calls
op.get_output(...), anyways, which calls imperative_run

---------

Co-authored-by: Allan Lin <allan.lin@centml.ai>
Co-authored-by: Yaoyao Ding <dingyaoyao.cs@gmail.com>