[Enhancements] add a vcude device to help mitigate compile time GPU memory usage (#302)
## The problem
During compilation, some passes needs to replace certain operators with
one or more new operators. For example, resolve_variants,
subgraph_rewrite. In doing so, hidet relies on `imperative_run` to
generate the new operators. This creates a lot of intermediate tensors
that can exceed the actual run-time memory consumption, which makes
larger models (such as Llama-7B) unable to compile even for a GPU with
24GB using fp16.

## Fix
We introduce a vcuda device that allows GPU tensors to be stored on CPU
and only transferred to GPU on demand. We call this v`cuda`. With this
change, any additional GPU memory usage during compilation is off-loaded
to CPU.

This might incur a bit compilation overhead when enabled, but given the
time-consuming nature of such large models, this compile time increase
is negligible.

Now, on RTX3090, this is the GPU memory consumption for running llama
test

```
Status of cuda:0 memory pool
   Allocated: 14081 MiB
        Peak: 14081 MiB
    Reserved: 1196 MiB
      Active: 12884 MiB
Status of cpu memory pool
   Allocated: 4 KiB
        Peak: 26022 MiB
    Reserved: 4 KiB
      Active: 0 Bytes
Status of vcuda:0 memory pool
   Allocated: 0 Bytes
        Peak: 25486 MiB
    Reserved: 0 Bytes
      Active: 0 Bytes
```

---------

Co-authored-by: Xin Li <xin@centml.ai>