[Compile Server] Add a new option to specify the cuda arch (#305)
Add the following option to specify the cuda target. 
```python
hidet.option.cuda.arch('sm_75')   # specify the target cuda arch
```

Originally, we will use the arch for the current gpu. However,
sometimes, we want to specify a different one, we can use the above
option to set the cuda arch.