[IR][Parser] Hidet IR grammar, parser and ir reconstructor (#354)
Currently, all IR roundtrips correctly in the test suite, for all
passes. The text is identical except for the attention ir module, which
has trivial changes (eg. (0 == wj) -> (wj == 0), or (!false) -> true).

Also fixed inline let stmt pass, so that types of tensors do not change
midway.

We need to add a dependency on lark, as the parser it generates is not
correct (does not have the same behavior as the original).

---------

Co-authored-by: Allan Lin <allan.lin@centml.ai>