[CI] Shut down on demand runners on failure (#64)
Previously, if a performance regression fails due to an exception, the
job that stops the runner VM instances will be skipped, leaving the
instances on. This will make the stop_instances job run even when
previous jobs failed. Not sure if always() will override the
inputs.shutdown_instances flag, if it does we can move it into the step
scope.