[PERF] Allow prologue fusion for `reduce` op  (#426)
Allow prologue for fp32 `reduce`. 

`reduce` uses vectorized calculations that don't allow to use fusing(it is possible but not implemented yet). For fp32 there are no vectors and we can enable fusion (with small modification `reduce` kernels itself).

Motivation.
In llama2 the part of the calculation is fp32 including `pow`+`reduce`. 

Performance improvement on llama2-7B +0.241%