[Graph][Ops] disable cublas matmul for parallel k (#431)
I found the stride calculation for the cublas batched matmul is
incorrect when we enable the parallel k optimization. And the wrong
strides lead to the mismatching results in Llama inference. It's a bit
complex to convert the matmuls in parallel k optimzation to a
canonicalized batched matmul in cublas, so I just disabled it.
Specifically, after splitting the K dimension, the extent in K dimension
for each partition might be unequal, and we have to check if the K
dimension is out-of-bound or not. But, actually, we didn't now. In
addition, this complicates the conversion to a normal batched matmul in
cublas.
closes #419

---------

Co-authored-by: Ubuntu <ubuntu@ip-172-31-43-134.us-east-2.compute.internal>
Co-authored-by: xiaocenxiaocen <xiao.zhang@centml.ai>