Fixed the format change on the new transformers version (#482)
### PR Comment:

In the new version of the `transformers` library (version 4.45.0) used
in CI, the `merges` field in the configuration has changed to a list of
lists. To ensure compatibility with this update, I have modified our
code base to accommodate this change.

Without this adjustment, the `test_tokenizer` in CI would fail to
execute successfully. This update ensures that the tests run as expected
with the new library version.