[CI] Shorten build-docs run time (#565)
Currently, the CI process takes nearly 3 hours to complete the
`build-docs` step. We suspect this is related to the recent change in
the `matmul` backend from `simt` to `mma`, which significantly expanded
the tuning space. To address this issue, we are adjusting the precision
from `float32` to `float16`. This change reduces the tuning complexity
and is expected to decrease compilation time, helping to optimize the CI
workflow. After this PR, the running time for `build-docs` should fall
back to 40 minutes.