Use one global cuda workspace for all the CompiledGraph (#603)
Fix the error in #564. When multiple CompileGraph exists, the cuda
workspace doesn't get reused. This causes accumulation of memory in LLMs
when all the subgraph exists at the same time.

This PR uses one global workspace for all the CompileGraph objects. This
reuses the memory.
Limitation: This PR assumes that we have one computation CUDA stream.