[BUG] Fix bugs in shared map implementation  (#608)
Shared map (used for implementation in-place ops) is more actual now
because `wait_tensor` from distributed must be in-place.

Note. This is a part of old haven't committed PR #73 that related to shared
map fixes only. In #73 there was also changes that relate to removing
identity ops.
For shared map fixed there were no concerns in #73. About removing
identity ops we decided to not commit them without support in fusion.