[BUG] fix attach hash to signature (#638)
Originally, this pass extracts all the kernels from the IR and adds
additional information to the name of the kernels. But we should update
all the places where the function variable is used. Previously, this
pass only updates the `LaunchKernelStmt`, which is only one place that
may use function variables. We should replace function variables in the
call expressions as well, which may also use the modified function
variables.

---------

Co-authored-by: xiaocenxiaocen <xiao.zhang@centml.ai>