add a flag for hexcute kernels (#642)
1. Add an option to enable experimental features. Usage:
```python
with hidet.option.context():
     hidet.option.hexcute_matmul(strategy='enable')
     ....
```
The valid values can be `enable`, `disable`, and `auto`. The `auto`
option will use a heuristic to determine whether to enable the Hexcute
kernel on the current GPUs. Now the heuristic is not implemented, and
will be implemented in the following PR.

---------

Co-authored-by: xiaocenxiaocen <xiao.zhang@centml.ai>
(cherry picked from commit bd5a327726adabfecfbf843e67b4c4a93ca4e837)