[Tests/CI] Update tests and add a temperary AMD CI (#693)
Previously, we only had official support for CUDA. Since the AMD GPUs
are getting more and more important and other packages like PyTorch,
vLLM, etc have better support for AMD GPUs, we are also going to support
AMD GPUs.

This PR tries to add the AMD CI as the first step. There are a fews
things that are related to this PR:
- Many existing PRs assumed we tested on a machine with NVIDIA GPUs
(directly use CUDA-specific features).
- Some tests have a parameterized `device` parameter to test over both
'cpu' and 'cuda' devices.

This PR introduces one automatic parameter (`device`) and several
markers (`requires_cuda`, `requires_cuda_hopper`, `requires_hip`) to
handle the cross-gpu tests. All existing tests have been updated to
follow the introduced convention.

There are three devices that are (going to be) supported by Hidet:
- NVIDIA GPUs (CUDA Platform)
- AMD GPUs (ROCm/HIP Platform)
- CPU

We need to test of all the model/operator tests on these devices.

To achieve so, we added a `device` fixture that will be filled with the
device that the current CI server supports.

For example, we could add a test like
```python
import pytest
import hidet
import hidet.testing

@pytest.mark.parameterize('n', [1, 10, 100])
def test_add(n, device):
    assert device in ['cpu', 'cuda', 'hip']
    a = hidet.randn([n], device=device)
    b = hidet.randn([n], device=device)
    c = hidet.ops.add(a, b)
    torch_c = a.torch() + b.torch()
    assert hidet.utils.assert_close(c, torch_c)
```

In the CUDA CI server, the device will be `cuda`; and in the HIP CI
server, the device will be `hip`.

See the `tests/conftest.py` for details where the `device` fixture is
defined.

Some tests are specific to some devices. For example, we might test some
primitive function for CUDA platform. For those
tests, we can use the `pytest.mark.requires_cuda` and
`pytest.mark.requires_hip` decorators to make sure that the test
is only run on the specific device. We rename the previuos
`pytest.mark.hopper` to `pytest.mark.requires_cuda_hopper` to indicate
that this test can only be run on an NVIDIA Hopper GPU.

For example, we could add a test like
```python
import pytest
import hidet
import hidet.testing

@pytest.mark.requires_cuda
def test_add_cuda():
    a = hidet.randn([10], device='cuda')
    b = hidet.randn([10], device='cuda')
    c = hidet.ops.add(a, b)
    torch_c = a.torch() + b.torch()
    assert hidet.utils.assert_close(c, torch_c)
```
that will be only run on the CUDA CI server.

Above docs are also put in `tests/README.md`.

1. Currently, we host the AMD GPU CI on a local server
with RX 7900 XTX and run all tests on that machine. In the future, we
should host the CI similar like CUDA in the cloud and paralleize the
tests under different sub-folders under `tests`.
2. Currently, the AMD GPU ci will only be triggered when the PR body
(and comments) contains a string "$ hidet-ci amdgpu". We could append
the following code snippet in the PRs that wants to test amdgpu:
```
$ hidet-ci amdgpu
```

(cherry picked from commit 25089bfa241ef9b76d83761c6b7d6b97febfcb80)