[HIP] Hip runtime - memory and device (#724)
This PR adds the runtime support of memory and device for HIP device.

This PR add the wrapper functions for the hip runtime functions in
`libamdhip64.so` and add 'hip' device.

The wrapper functions around `libamdhip64.so` will be **replaced** by
the `hip-python` pakcage later, thus the wrapper code in this PR will be
removed later.

Since the ci does not have the `libamdhip64.so` library, I crafted a
dummy class to replace it to pass the tests on CUDA machine. The whole
ffi module will be removed thus it should be okay.

---------

Co-authored-by: yudi0201 <82490690+yudi0201@users.noreply.github.com>
(cherry picked from commit 2c11f6fb37a92a1a6cfcea80fc27d94aa2bc0892)