[BUG] fix flatten tensor index pass (#710)
This PR fixes an issue of the `flatten_tensor_index_pass`. This pass
will flatten all the variables that are multi-dimensional tensors. It
will update the parameter's type in the function, so we should also
update the call expression where the function is used. We maintain a
table of all the functions whose parameter types have been updated,
ensuring we know which function variable should be updated
correspondingly. But this requires that we visit the functions in
callee-first order, which means the callee should be visited first.
Previously, the pass doesn't guarantee the visit order of the functions,
which could potentially cause errros. This PR fixes this issue.

---------

Co-authored-by: xiao zhang <xiao.zhang@centml.ai>
(cherry picked from commit dd78d4eb96e674d464dcf96613ed2455de37d1a8)