[Hopper] add a cost model (#678)
Currently, it is not used in our code base.

Before Hopper:

```python
for ko in range(cdiv(K, BK)):
     copy(gA[:, :, ko + prefetch_stage], sA)
     copy(gB[:, :, ko + prefetch_stage], sB)

     for ki in range(BK // k_tile):
         copy(sA[:, :, ki], rA)
         copy(sB[:, :, ki], rB)     # first load from shared memory
         mma(rA, rB, rC)          # second consumed by `mma` instruction
```
The swizzling layout can be determined by maximizing the throughput of
`copy(sB[:, :, ki], rB)`.

After Hopper:

```python
for ko in range(cdiv(K, BK)):
     copy(gA[:, :, ko + prefetch_stage], sA)              (S1)
     copy(gB[:, :, ko + prefetch_stage], sB)              (S2)

     for ki in range(BK // k_tile):
         copy(sA[:, :, ki], rA)
         wgmma(rA, sB[:, :, ki], rC)          # fuse the first and second steps on the hardware.
```
When the `wgmma` is used, the first and the second steps are fused. The
hardware provides four layouts for the shared memory layout B, which are
1. `INTERLEAVED`: means without swizzling
2. `SW32`: the size of the smallest swizzling component is 32 bytes
3. `SW64`: the size of the smallest swizzling component is 64 bytes
4. `SW128`: the size of the smallest swizzling component is 128 bytes

The four layouts are all valid for the `wgmma` instruction, so we cannot
choose the best one from the local information (the information of the
`wgmma` instruction itself). However, the selection of the layouts will
impact the performance of the `copy` operations in `S2`. It will
influence which vector load/store instruction can be used to perform the
data movement and whether the copy can be coalesced. The cost model
allows us to generate the four variants and select the most efficient
one with global information.

Yes, we can. If we annotate the layout of `sB`, `SW32` for example, the
compiler can only find one solution, then the cost model will not be
used. So, the performance is deterministic.

If we want the compiler to determine the best layouts automatically, we
need a cost model to evaluate which solution is the best. The cost model
is for the completeness of the automatic layout inference system. We can
disable the cost model by annotating the layout information to get the
deterministic performance.

---------

Co-authored-by: xiaocenxiaocen <xiao.zhang@centml.ai>
(cherry picked from commit 25f036cdc8d69ef2d334000b829dbdaeb0eeff1e)