[Codegen][Runtime] Add try-catch to protect the public function (#871)
When we launch a kernel in public function, we might throw
HidetExcpetion in the C++. But our public function did not catch it.
This makes python to stop at the C and the execption will not propogate
to python. Similar like our runtime API function, we should add a
try-catch to gurad the public fuctions.

This PR also updates the Codegen to avoiding generating some redundant
parathesis. Like ((((a + b) + c) + d) + e) => (a + b + c + d + e). They
make the analysis on the cuda code hard when we debug on the CUDA code.
We have use '(' and ')' to protect all the expressions with computation
priority lower than addition, thus it's safe.

(cherry picked from commit d5ca712ef5b763d7a9f2426c195b96b05484f459)