[Pass] Optimization for addition chain  (#874)
This PR introduces an optimization pass to perform optimizations on
addition chain: e1 + e2 + e3 ... + en.

Our existing rule-based simplifier only the addition with two operands.
It can not handle the case when two operands are not in the same Add
expression. For example, we encounter the following kind of expressions:
```
a: int32
b: int32

c = (a * 10 / 10 + b) + a % 10
```
Even we have rules like
```
a * C / C + a % C => a
```
we will not apply the optimzation on c since `a * C / C` are not in an
Add with `a % C`.

This PR handle this kind of optimization by considering the addition
chain in a whole. The downside is that we need to manually do the
pattern matching and replacing.

Example (in the test):
```python
  with hidet.script_module() as script_module:

      @hidet.script
      def func(a: int32, b: int32):
          attrs.func_kind = 'cuda_kernel'
          c = a / 10 * 10 + b + a % 10

  module = script_module.ir_module()
  print(module)
  """
  def func(
      a: int32,
      b: int32
  )
      # kind: cuda_kernel
      # func_kind: cuda_kernel
      declare c: int32 = ((((a / 10) * 10) + b) + (a % 10))
  """
  transform = simplify_addition_chain_pass()
  transformed_module = transform(module)
  print(transformed_module)
  """
  def func(
      a: int32,
      b: int32
  )
      # kind: cuda_kernel
      # func_kind: cuda_kernel
      declare c: int32 = (a + b)
  """

  assert "declare c: int32 = (a + b)" in str(transformed_module)
```

This PR also fixed a bug in Hidet Script when we directly launch a
kernel from public function, previously, the launch arguments are not
replaced with the variables defined in the host function.

(cherry picked from commit 080066e3e3767cc2cd46c2994482f83efe5cd77a)