[Package] Refactor dependency configuration (#900)
We currently have two places to store the dependencies of hidet:
- `requirements.txt` and `requirements-dev.txt`
- pyproject.toml

This PR migrate the `requirements.txt` and `requirements-dev.txt` into
`pyproject.toml` so that we keep one ground truth for the dependencies
and dev dependencies.

Others related to dependencies:
1. add `cmake` to the `build-system` `requires` section in
`pyproject.toml` since our runtime library requires `cmake` to work
build.
2. remove the `astunparse` dependency (it's a 3rdparty library on pypi)
and use the standard library function `ast.unparse` instead. Python 3.9
introduces this function and hidet requires python 3.9+ now.

[04/11] Update:
This PR also phase out "model transolator". This is an effort from one
of our preivuosly interns that translates a pytorch model (in
python code) to our own model definition with hidet.graph.nn.Module.
Both idea and technique used are cool. But it's not actually finished or
used, due to its highly complexity and a large amount of features to
support. With torch.compile, we rely on dynamo to provide fx.Graph and
it's time to phase out this utility. It depends on `astunparse` so I
notice it.

(cherry picked from commit 7234e0aca2f56a2723568c5d2dedb0abc3470996)