[Bug] Release the reserved memory in hidet for kv cache (#920)
When creating a workspace for compiled graph, we eventually select the
largest workspace and reuse it for all graphs. But we don't release the
memory of workspaces that we don't need anymore. This PR fixes that.

Co-authored-by: Zhumakhan <nazirzhumakhan@gmail,.com>
(cherry picked from commit a58e82c05595d200b60ca478db73fab5e4af921c)