[Feature][CuTe] add `mbarrier` operators in Hexcute (#938)
This PR is the prerequesite of `TMA` support in Hexcute.

# Changes
1. refactor some layout utilities in CuTe.
2. add mbarrier related ops.
3. add a pass to determine the number of the threads that participate in
the barrier operation
4. code generation for mbarrier related ops

# Example
TMA relies on a mbarrier object to notify other threads the completion
of the TMA. After this change, the copy operation can accept a mbarrier
obj.
```
copy(tiled_copy, src_tensor, dst_tensor, mbarrier=mbar)
```

# Design
1. The user can choose passing the mbarrier object or not. If not
providing the mbarrier, the compiler won't select the TMA instruction to
perform the data movement.
2. With mbarrier provided, the compiler will try to determine layouts
for the TMA copy instruction. If the compiler cannot find a layout for
the TMA, it will fallback to normal copy instruction. In such cases, the
mbarrier will still be used to indicate the completion of the copy
operation (for asynchronous programming model).

---------

Co-authored-by: xiao zhang <xiao.zhang@centml.ai>
(cherry picked from commit b68c605762f0ac82b7d9d9b1a347ecc958a887be)