[Bugfix] Add signal handler to clean up NCCL sync file (#955)
fix #950

We use local file system to broadcast NCCL's Unique ID to all the
processes in the group. The file we use won't get cleaned when we kill
the process with SIGINT(Ctrl+C) or SIGTERM(kill -15). On the next run,
if any rank other than rank 0 read this file first, it will directly
read from the end to get the previous Unique ID instead of creating a
new one. Only rank 0 creates a new Unique ID. So there will be at least
one rank (that is rank 0) that is in possess of a different new Unique
ID. This will cause `ncclCommInitRank` to fail due to incorrect Unique
ID.

But if we send `kill -9`, the file will still be there. In that case we
have to manually remove the file (We can't remove the file in
multiprocess scenario as we don't have a way to sync all the processes
in hidet before creating a process group).

(cherry picked from commit 99aa56d906d5cbd5be55a28884521edaf3f2c5d6)