[Apps] Phase out the app-level abstraction (#971)
We do not plan to push the support on the app level (a level contains
multiple CompliledGraph and support the execution of an application like
LLM serving). To reduce the maintainancy burden, we decide to deprecate
this part of hidet.

This PR phases out the support to make the codebase more clean and
focus.

(cherry picked from commit e8d709f2d27495adef41bcec6e80fac9ac79960e)