[Version] Use `setuptools-scm` to manage hidet version (#988)
Currently, we manually manage the hidet version by hardcoding it in the
`pyproject.toml` and `hidet/version.py`.


There are several disadvantages:
1. For nightly built hidet, we can not track the commit we used to build
the wheel.
2. We need to manually create a PR to update the version.

Packages like vllm use `setuptools_scm` to manage their version. 
(https://github.com/vllm-project/vllm/blob/main/pyproject.toml#L35)

How it works: in short, it will determine the version according to the
git tag of the commit history:
- if the commit has tag like v0.5.1, the hidet version will be `v0.5.1`.
- if the latest ancestor has tag v0.5.1, then the version will be
`v0.5.1.dev32+g480eea3.d20250526`, which includes the number of commits
on top of the tag v0.5.1 (e.g., 32), the commit sha (480eea3), and the
date of the commit.
See: https://setuptools-scm.readthedocs.io/en/latest/ and
https://setuptools-scm.readthedocs.io/en/latest/usage/#default-versioning-scheme

This PR enables `setuptools-scm` in hidet.

Other small changes:
1. Remove the manuall installation of cmake, which is specified in
`[build-system] require` section in `pyproject.toml`.
2. Remove the `build-wheel` job in `tests.yaml`, since we directly use
the package in the docker image.
3. Let the `build-wheel.sh` directly build the package under the repo,
instead of copying the files into subfolder to perform the building,
which make sure `setuptools-scm` to recognize that we are building
against a commit without modification, thus not adding suffix like
`d20250526`.