[PERF] [hidet+dmwl] Add einops.rearrange support in hidet (#961)
Qwen2.5-vl has a lot of einops.rearrange. Currently, we do not support
this and use inductor's `rearrange` and return to hidet after.

This causes overhead from entering and exiting hidet and we also cannot
optimize the rearrange with other ops.

einops.rearrange maps directly to hidet.ops.rearrange but
einops.rearrange is human readable while hidet.ops.rearrange uses a List
of Lists to represent a plan.

This PR parses the human readable rearrange plan and maps it to the
hidet plan and overrides the einops.rearrange op

Issue #936

---------

Signed-off-by: Peterson Guo <peterson.guo@centml.ai>