[Primitive] Add volatile mark for cp-async instructions (#1018)
Though we did not find correctness issues in existing generated kernels,
but it's safer to add volatile to cp-async instructions. It prevents
nvcc to reorder these instructions. I heard that nvcc sometimes will
reorder instructions to make `cp.async.commit` and `cp.async.wait` stay
together, which is not desired.