[PERF] [llama8B] Add subgraph rewrite rule to fuse Matmul and cast (#1021)
Unfused Graph:

![image](https://github.com/user-attachments/assets/766e5b8d-3882-400d-a398-e513046e5a63)

Original Behaviour:

![image](https://github.com/user-attachments/assets/67c52eed-145b-4cdf-afb3-6ffa492a29d5)

New Behaviour:

![image](https://github.com/user-attachments/assets/8a3b0588-9548-45f2-a0e9-8329a39c8533)

Originally, the CastOp was being left out of the fusion because after a
epilogue fusion pass, everything after matmul would be fused and CastOp
would be left as a prologue operator. Because prologue fusion is
disabled on Matmul, CastOp would never be fused.

Fixed this issue by adding a MultiInputCompositeElementWiseOp which
combines 2 UnaryOps with different inputs and 1 BinaryElementWiseOp,
allowing this shape to be fused with Matmul.